feat: Landing page
setup for landing page
