+++ apps/nuxt3-ssr/components/layouts/LandingPage.vue
@@ -1,0 +1,9 @@
+<script setup>
+import Container from "./../Container.vue";
+</script>
+
+<template>
+  <Container>
+    <slot></slot>
+  </Container>
+</template>
